Use solr user instead of root user.
This commit adds a solr user and invokes the solr service with the user.
I did not implement this for the centos initd because I do not have time
to test it for centos.  Although the user has been changed the centos
version should run the same as always as that version is ran as the root
user.